Added project for DAL using Entity Framework and moved DAL interfaces and classes to separate projects
