+++ src/CookBook/CookBook.Api.BL/CookBook.Api.BL.csproj
@@ -16,9 +16,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\CookBook.Api.DAL.Common\CookBook.Api.DAL.Common.csproj" />
     <ProjectReference Include="..\CookBook.Common.BL\CookBook.Common.BL.csproj" />
     <ProjectReference Include="..\CookBook.Common.Models\CookBook.Common.Models.csproj" />
-    <ProjectReference Include="..\CookBook.Api.DAL\CookBook.Api.DAL.csproj" />
+    <ProjectReference Include="..\CookBook.Api.DAL\CookBook.Api.DAL.Memory.csproj" />
   </ItemGroup>
 
 </Project>
