+++ src/CookBook/CookBook.Api.BL/Facades/IngredientFacade.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using AutoMapper;
-using CookBook.Api.DAL.Entities;
-using CookBook.Api.DAL.Repositories;
+using CookBook.Api.DAL.Common.Entities;
+using CookBook.Api.DAL.Common.Repositories;
 using CookBook.Common.BL.Facades;
 using CookBook.Common.Models;
