+++ src/CookBook/CookBook.Api.DAL.Common/Entities/IngredientAmountEntity.cs
@@ -2,7 +2,7 @@
 using AutoMapper;
 using CookBook.Common.Enums;
 
-namespace CookBook.Api.DAL.Entities
+namespace CookBook.Api.DAL.Common.Entities
 {
     public record IngredientAmountEntity : EntityBase
     {
