+++ src/CookBook/CookBook.Api.DAL.Common/Entities/IngredientEntity.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using AutoMapper;
 
-namespace CookBook.Api.DAL.Entities
+namespace CookBook.Api.DAL.Common.Entities
 {
     public record IngredientEntity : EntityBase
     {
