+++ src/CookBook/CookBook.Api.DAL.Common/Repositories/IApiRepository.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using CookBook.Api.DAL.Entities;
+using CookBook.Api.DAL.Common.Entities.Interfaces;
 
-namespace CookBook.Api.DAL.Repositories
+namespace CookBook.Api.DAL.Common.Repositories
 {
     public interface IApiRepository<TEntity>
         where TEntity : IEntity
