+++ src/CookBook/CookBook.Api.DAL.Common/Repositories/IRecipeRepository.cs
@@ -1,0 +1,8 @@
+﻿using CookBook.Api.DAL.Common.Entities;
+
+namespace CookBook.Api.DAL.Common.Repositories
+{
+    public interface IRecipeRepository : IApiRepository<RecipeEntity>
+    {
+    }
+}
