+++ src/CookBook/CookBook.Api.DAL.Memory/CookBook.Api.DAL.Memory.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\CookBook.Common\CookBook.Common.csproj" />
+    <ProjectReference Include="..\CookBook.Api.DAL.Common\CookBook.Api.DAL.Common.csproj" />
   </ItemGroup>
 
   <ItemGroup>
