+++ src/CookBook/CookBook.Api.DAL.Memory/CookBook.Api.DAL.Memory.csproj.DotSettings
