--- src/CookBook/CookBook.Api.DAL/Repositories/Interfaces/IRecipeRepository.cs
@@ -1,8 +1,0 @@
-﻿using CookBook.Api.DAL.Entities;
-
-namespace CookBook.Api.DAL.Repositories
-{
-    public interface IRecipeRepository : IApiRepository<RecipeEntity>
-    {
-    }
-}
