+++ src/CookBook/CookBook.sln
@@ -25,12 +25,18 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CookBook.Web.BL", "CookBook
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CookBook.Web.DAL", "CookBook.Web.DAL\CookBook.Web.DAL.csproj", "{B0D7D099-6E7E-4476-A8FE-3573E5857BB6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CookBook.Api.DAL", "CookBook.Api.DAL\CookBook.Api.DAL.csproj", "{E914A67E-C3AC-45FB-9EDB-7062AE1C6E90}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CookBook.Api.DAL.Memory", "CookBook.Api.DAL.Memory\CookBook.Api.DAL.Memory.csproj", "{E914A67E-C3AC-45FB-9EDB-7062AE1C6E90}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{1B830E23-DE7F-451B-B5CC-EB70A66AD6BB}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DAL", "DAL", "{DA7BE6FF-5E94-4000-9BD7-A7DA82EF224F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CookBook.Api.DAL.Common", "CookBook.Api.DAL.Common\CookBook.Api.DAL.Common.csproj", "{EB5E8EB5-7F0B-4BDB-8045-617718084B46}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CookBook.Api.DAL.EF", "CookBook.Api.DAL.EF\CookBook.Api.DAL.EF.csproj", "{EC6D7327-9903-4738-B5F4-ECAF8F993AA0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -74,6 +80,14 @@ Global
 		{E914A67E-C3AC-45FB-9EDB-7062AE1C6E90}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E914A67E-C3AC-45FB-9EDB-7062AE1C6E90}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E914A67E-C3AC-45FB-9EDB-7062AE1C6E90}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EB5E8EB5-7F0B-4BDB-8045-617718084B46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EB5E8EB5-7F0B-4BDB-8045-617718084B46}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EB5E8EB5-7F0B-4BDB-8045-617718084B46}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EB5E8EB5-7F0B-4BDB-8045-617718084B46}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EC6D7327-9903-4738-B5F4-ECAF8F993AA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EC6D7327-9903-4738-B5F4-ECAF8F993AA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EC6D7327-9903-4738-B5F4-ECAF8F993AA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EC6D7327-9903-4738-B5F4-ECAF8F993AA0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -87,7 +101,10 @@ Global
 		{AE669C4D-FC6C-4E2F-8722-1D6FCE6FB0FD} = {777F05CE-DC73-46AE-BA0F-CCF368D381A1}
 		{0CCE7288-4F07-4F5A-8160-ACC5E7988DF2} = {777F05CE-DC73-46AE-BA0F-CCF368D381A1}
 		{B0D7D099-6E7E-4476-A8FE-3573E5857BB6} = {777F05CE-DC73-46AE-BA0F-CCF368D381A1}
-		{E914A67E-C3AC-45FB-9EDB-7062AE1C6E90} = {CE4F5F83-8F08-4F26-A46D-52370701657D}
+		{E914A67E-C3AC-45FB-9EDB-7062AE1C6E90} = {DA7BE6FF-5E94-4000-9BD7-A7DA82EF224F}
+		{DA7BE6FF-5E94-4000-9BD7-A7DA82EF224F} = {CE4F5F83-8F08-4F26-A46D-52370701657D}
+		{EB5E8EB5-7F0B-4BDB-8045-617718084B46} = {DA7BE6FF-5E94-4000-9BD7-A7DA82EF224F}
+		{EC6D7327-9903-4738-B5F4-ECAF8F993AA0} = {DA7BE6FF-5E94-4000-9BD7-A7DA82EF224F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {10E5D329-1983-4C74-838B-CA0CFC4A2256}
